+++ src/index.core.ts
@@ -31,6 +31,7 @@ export {
   AdtContentTypesBase,
   AdtContentTypesModern,
 } from './core/shared/contentTypes';
+export { changeTransportTaskType } from './core/transport/objects';
 export { parseCreatedTransport } from './core/transport/parseCreatedTransport';
 export { parseObjectEntries } from './core/transport/parseObjectEntries';
 export { parseTransportTree } from './core/transport/parseTransportTree';